+++ dev_scripts_helpers/dockerize/dockerized_template.py
@@ -3,15 +3,13 @@
 """
 Dockerized template.
 
-This script is a template for creating a Dockerized script. It is
-intended as a template to explain the process.
+This script is a template for creating a Dockerized script.
 """
 
 import argparse
 import logging
 
 import helpers.hdbg as hdbg
-import helpers.hdocker as hdocker
 import helpers.hparser as hparser
 
 _LOG = logging.getLogger(__name__)
@@ -23,7 +21,7 @@ def _parse() -> argparse.ArgumentParser:
         description=__doc__,
         formatter_class=argparse.RawDescriptionHelpFormatter,
     )
-    # FILL THIS.
+    # TODO(*): Add more options.
     # parser.add_argument(
     #     "--docx_file",
     #     required=True,
@@ -33,6 +31,7 @@ def _parse() -> argparse.ArgumentParser:
     # Add Docker-specific arguments (e.g., --dockerized_force_rebuild,
     # --dockerized_use_sudo).
     hparser.add_dockerized_script_arg(parser)
+    # Add logging verbosity parsing.
     hparser.add_verbosity_arg(parser)
     return parser
 
@@ -42,16 +41,16 @@ def _main(parser: argparse.ArgumentParser) -> None:
     hdbg.init_logger(
         verbosity=args.log_level, use_exec_path=True, force_white=False
     )
-    # FILL THIS.
-    pandoc_cmd = ()
-    _LOG.debug("Command: %s", pandoc_cmd)
-    hdocker.run_dockerized_pandoc(
-        pandoc_cmd,
-        container_type="pandoc_only",
-        force_rebuild=args.dockerized_force_rebuild,
-        use_sudo=args.dockerized_use_sudo,
-    )
-    _LOG.info("Finished converting '%s' to '%s'.", args.docx_file, args.md_file)
+    # TODO(*): Implement this.
+    # pandoc_cmd = ()
+    # _LOG.debug("Command: %s", pandoc_cmd)
+    # hdocker.run_dockerized_pandoc(
+    #    pandoc_cmd,
+    #    container_type="pandoc_only",
+    #    force_rebuild=args.dockerized_force_rebuild,
+    #    use_sudo=args.dockerized_use_sudo,
+    # )
+    # _LOG.info("Finished converting '%s' to '%s'.", args.docx_file, args.md_file)
 
 
 if __name__ == "__main__":
