+++ dev_scripts_helpers/documentation/dockerized_graphviz.py
@@ -7,7 +7,7 @@ import argparse
 import logging
 
 import helpers.hdbg as hdbg
-import helpers.hdocker as hdocker
+import helpers.hdockerized_executables as hdocexec
 import helpers.hparser as hparser
 
 _LOG = logging.getLogger(__name__)
@@ -34,7 +34,7 @@ def _main(parser: argparse.ArgumentParser) -> None:
     hdbg.init_logger(
         verbosity=args.log_level, use_exec_path=True, force_white=False
     )
-    hdocker.run_dockerized_graphviz(
+    hdocexec.run_dockerized_graphviz(
         args.input,
         cmd_opts,
         args.output,
