+++ dev_scripts_helpers/documentation/dockerized_tikz_to_bitmap.py
@@ -8,7 +8,7 @@ import argparse
 import logging
 
 import helpers.hdbg as hdbg
-import helpers.hdocker as hdocker
+import helpers.hdockerized_executables as hdocexec
 import helpers.hparser as hparser
 
 _LOG = logging.getLogger(__name__)
@@ -37,7 +37,7 @@ def _main(parser: argparse.ArgumentParser) -> None:
     hdbg.init_logger(
         verbosity=args.log_level, use_exec_path=True, force_white=False
     )
-    hdocker.run_dockerized_tikz_to_bitmap(
+    hdocexec.run_dockerized_tikz_to_bitmap(
         args.input,
         cmd_opts,
         args.output,
