+++ dev_scripts_helpers/documentation/preprocess_notes.py
@@ -181,13 +181,13 @@ def _transform_lines(txt: str, type_: str, *, is_qa: bool = False) -> str:
         # 7) Process question.
         if _TRACE:
             _LOG.debug("# Process question.")
-        # if type_ == "slides":
-        #     do_continue, line = _process_question_to_slides(line)
-        # else:
-        #     do_continue, line = _process_question_to_markdown(line)
-        # if do_continue:
-        #     out.append(line)
-        #     continue
+        if type_ == "slides":
+            do_continue, line = _process_question_to_slides(line)
+        else:
+            do_continue, line = _process_question_to_markdown(line)
+        if do_continue:
+            out.append(line)
+            continue
         # 8) Process empty lines in the questions and answers.
         if _TRACE:
             _LOG.debug("# Process empty lines in the questions and answers.")
@@ -227,38 +227,6 @@ def _transform_lines(txt: str, type_: str, *, is_qa: bool = False) -> str:
                     or next_line_is_verbatim
                 ):
                     out.append(" " * _NUM_SPACES + line)
-    #
-    if type_ == "slides":
-
-        def _transform(slide_text: List[str]) -> str:
-            slide_text = "\n".join(slide_text)
-            if not hmarkdo.has_color_command(slide_text):
-                text_out = hmarkdo.colorize_bullet_points_in_slide(
-                    slide_text, use_abbreviations=False
-                )
-            else:
-                text_out = slide_text
-            text_out = text_out.split("\n")
-            return text_out
-
-        out = "\n".join(out)
-        out = hmarkdo.process_slides(out, _transform)
-        out = out.split("\n")
-
-    # out = out.split("\n")
-    out_tmp = []
-    for line in out:
-        if type_ == "slides":
-            do_continue, line = _process_question_to_slides(line)
-        else:
-            do_continue, line = _process_question_to_markdown(line)
-        if do_continue:
-            out_tmp.append(line)
-            continue
-        out_tmp.append(line)
-    out = out_tmp
-    # out = "\n".join(out_tmp)
-
     # c) Clean up.
     _LOG.debug("Clean up")
     # Remove all the lines with only spaces.
