+++ dev_scripts_helpers/documentation/test/outcomes/Test_lint_notes2.test_process1/output/test.txt
@@ -1,4 +1,4 @@
-* Gradient descent for logistic regression
+* Gradient Descent for Logistic Regression
 - The typical implementations of gradient descent (basic or advanced) need two
   inputs:
   - The cost function $E_{in}(\vw)$ (to monitor convergence)
