+++ dev_scripts_helpers/documentation/test/test_lint_notes.py
@@ -6,7 +6,7 @@ import pytest
 
 import dev_scripts_helpers.documentation.lint_notes as dshdlino
 import helpers.hdbg as hdbg
-import helpers.hdocker as hdocker
+import helpers.hdockerized_executables as hdocexec
 import helpers.hgit as hgit
 import helpers.hio as hio
 import helpers.hprint as hprint
@@ -158,7 +158,7 @@ class Test_lint_notes2(hunitest.TestCase):
             -   avoid non-essential tasks
         """
         exp = r"""
-        * Good time management
+        * Good Time Management
 
         1. Choose the right tasks
            - Avoid non-essential tasks
@@ -190,7 +190,7 @@ class Test_lint_notes2(hunitest.TestCase):
 
         # Good
 
-        - Good time management
+        - Good Time Management
           1. Choose the right tasks
           - Avoid non-essential tasks
 
@@ -243,15 +243,15 @@ class Test_lint_notes2(hunitest.TestCase):
         For some reason prettier replaces - with * when there are 2 empty lines.
         """
         txt = self._get_text_problematic_for_prettier1()
-        act = hdocker.prettier_on_str(txt, file_type="txt")
+        act = hdocexec.prettier_on_str(txt, file_type="txt")
         exp = r"""
-        - Python formatting
+        - Python Formatting
 
         * Python has several built-in ways of formatting strings
           1. `%` format operator
           2. `format` and `str.format`
 
-        - `%` format operator
+        - `%` Format Operator
 
         * Text template as a format string
           - Values to insert are provided as a value or a `tuple`
@@ -265,13 +265,12 @@ class Test_lint_notes2(hunitest.TestCase):
         """
         txt = self._get_text_problematic_for_prettier1()
         exp = r"""
-        * Python formatting
+        * Python Formatting
         - Python has several built-in ways of formatting strings
-
           1. `%` format operator
           2. `format` and `str.format`
 
-        * `%` format operator
+        * `%` Format Operator
         - Text template as a format string
           - Values to insert are provided as a value or a `tuple`
         """
@@ -462,41 +461,3 @@ class Test_lint_notes_cmd_line1(hunitest.TestCase):
         output_txt = self.run_lint_notes(in_file, type_, cmd_opts)
         # Check.
         self.check_string(output_txt)
-
-
-# #############################################################################
-# Test_improve_header_and_slide_titles1
-# #############################################################################
-
-
-class Test_improve_header_and_slide_titles1(hunitest.TestCase):
-    """
-    Test the function `_improve_header_and_slide_titles`.
-    """
-
-    def helper(self, txt: str, exp: str) -> None:
-        txt = hprint.dedent(txt)
-        exp = hprint.dedent(exp)
-        act = dshdlino._improve_header_and_slide_titles(txt)
-        self.assert_equal(act, exp)
-
-    def test1(self) -> None:
-        txt = r"""
-        * ML theory
-        """
-        exp = r"""
-        * ML Theory
-        """
-        self.helper(txt, exp)
-
-    def test2(self) -> None:
-        """
-        Test the function `_improve_header_and_slide_titles`.
-        """
-        txt = r"""
-        * A map of machine learning
-        """
-        exp = r"""
-        * A Map of Machine Learning
-        """
-        self.helper(txt, exp)
